refactor(DC-704): extract rate limiter setup into RateLimitingSetup
