+++ apps/portal/src/SEBT.Portal.Api/SEBT.Portal.Api.csproj
@@ -132,7 +132,6 @@
     <Folder Include="plugins-dc\" />
     <Folder Include="plugins-co\" />
     <Folder Include="Startup\Services\" />
-    <Folder Include="Startup\Setup\" />
   </ItemGroup>
 
 </Project>
